+++ Source/SwiftLintBuiltInRules/Models/BuiltInRules.swift
@@ -204,6 +204,7 @@ public let builtInRules: [Rule.Type] = [
     TypesafeArrayInitRule.self,
     UnavailableConditionRule.self,
     UnavailableFunctionRule.self,
+    UnhandledThrowingTaskRule.self,
     UnneededBreakInSwitchRule.self,
     UnneededParenthesesInClosureArgumentRule.self,
     UnownedVariableCaptureRule.self,